fix: update root page to be signin
