+++ app/(web-client)/user/[userId]/worker/calendar/page.tsx
@@ -114,12 +114,12 @@ const WorkerCalendarPage = () => {
     }
 
     if (!user) {
-      router.push(`/signin?redirect=${pathname}`);
+      router.push(`/?redirect=${pathname}`);
       return;
     }
 
     if (authUserId !== pageUserId) {
-      router.push(`/signin?error=unauthorized`);
+      router.push(`/?error=unauthorized`);
       return;
     }
