+++ app/components/onboarding/VideoRecorderBubble.module.css
@@ -56,6 +56,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  position: relative
 }
