+++ app/globals.css
@@ -1,4 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100..900;1,100..900&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Source+Sans+3:ital,wght@0,200..900;1,200..900&display=swap');
 :root {
   --background: #ffffff;
   --foreground: #161616;
@@ -26,8 +27,7 @@
   --success-color: #10b981;
   --border-color: #e5e7eb;
   --border-radius: 0.375rem; /* 6px */
-  --font-family: 'Roboto',
-    "Helvetica Neue", Arial, sans-serif; /* Added Source Sans Pro */
+  --font-family: "Roboto", sans-serif;
 
   /* Added Buyer Accent Colors */
   --buyer-accent-color: #22d3ee; /* Cyan/Aqua */
@@ -52,7 +52,7 @@
 }
 
 body {
-  /* font-family: var(--font-family); */
+  font-family: var(--font-family);
   color: var(--text-color);
   background-color: #161616; /* Dark background to match the app theme */
   margin: 0;
