--- app/page.module.css
@@ -1,311 +1,0 @@
-.page {
-  --gray-rgb: 0, 0, 0;
-  --gray-alpha-200: rgba(var(--gray-rgb), 0.08);
-  --gray-alpha-100: rgba(var(--gray-rgb), 0.05);
-
-  --primary-color: #0070f3;
-  --primary-hover: #0051a2;
-  --secondary-color: #666;
-  --success-color: #10b981;
-  --warning-color: #f59e0b;
-  --planned-color: #6b7280;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  min-height: 100svh;
-  padding: 40px 20px;
-  font-family: var(--font-geist-sans);
-}
-
-@media (prefers-color-scheme: dark) {
-  .page {
-    --gray-rgb: 255, 255, 255;
-    --gray-alpha-200: rgba(var(--gray-rgb), 0.145);
-    --gray-alpha-100: rgba(var(--gray-rgb), 0.06);
-
-    --button-primary-hover: #ccc;
-    --button-secondary-hover: #1a1a1a;
-  }
-}
-
-.main {
-  width: 100%;
-  max-width: 900px;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-  grid-row-start: 2;
-}
-
-.main ol {
-  font-family: var(--font-geist-mono);
-  padding-left: 0;
-  margin: 0;
-  font-size: 14px;
-  line-height: 24px;
-  letter-spacing: -0.01em;
-  list-style-position: inside;
-}
-
-.main li:not(:last-of-type) {
-  margin-bottom: 8px;
-}
-
-.main code {
-  font-family: inherit;
-  background: var(--gray-alpha-100);
-  padding: 2px 4px;
-  border-radius: 4px;
-  font-weight: 600;
-}
-
-.summary {
-  cursor: pointer;
-  padding: 1rem;
-  background-color: var(--gray-alpha-100);
-  border-radius: 8px;
-  font-weight: 500;
-}
-
-.projectInfo {
-  padding: 1.5rem;
-  background-color: var(--gray-alpha-100);
-  border-radius: 0 0 8px 8px;
-  margin-top: -8px;
-}
-
-.projectInfo h2 {
-  margin-top: 0;
-  margin-bottom: 1rem;
-  font-size: 1.5rem;
-}
-
-.projectInfo ul {
-  list-style-type: disc;
-  margin-left: 1.5rem;
-}
-
-.userSection {
-  background-color: var(--gray-alpha-100);
-  padding: 1.5rem;
-  border-radius: 8px;
-}
-
-.userInfo {
-  margin-bottom: 1rem;
-}
-
-.userActions {
-  display: flex;
-  gap: 1rem;
-  flex-wrap: wrap;
-}
-
-.primary {
-  background-color: var(--primary-color);
-  color: white;
-  padding: 0.5rem 1rem;
-  border-radius: 6px;
-  text-decoration: none;
-  font-weight: 500;
-  border: none;
-  cursor: pointer;
-}
-
-.primary:hover {
-  background-color: var(--primary-hover);
-  text-decoration: none;
-}
-
-.secondary {
-  background-color: transparent;
-  color: var(--secondary-color);
-  padding: 0.5rem 1rem;
-  border-radius: 6px;
-  text-decoration: none;
-  font-weight: 500;
-  border: 1px solid var(--secondary-color);
-  cursor: pointer;
-}
-
-.secondary:hover {
-  background-color: var(--gray-alpha-100);
-  text-decoration: none;
-}
-
-.pagesSection {
-  background-color: var(--gray-alpha-100);
-  padding: 1.5rem;
-  border-radius: 8px;
-  width: 100%;
-}
-
-.pagesList {
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-}
-
-.pageItem {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0.75rem;
-  background-color: white;
-  border-radius: 6px;
-}
-
-.pageName {
-  font-family: var(--font-geist-mono);
-  color: var(--secondary-color);
-}
-
-.badge {
-  padding: 0.25rem 0.75rem;
-  border-radius: 999px;
-  font-size: 0.875rem;
-  font-weight: 500;
-}
-
-.badge[data-status="complete"] {
-  background-color: var(--success-color);
-  color: white;
-}
-
-.badge[data-status="in-progress"] {
-  background-color: var(--warning-color);
-  color: white;
-}
-
-.badge[data-status="planned"] {
-  background-color: var(--planned-color);
-  color: white;
-}
-
-.ctas {
-  display: flex;
-  gap: 16px;
-}
-
-.ctas a {
-  appearance: none;
-  border-radius: 128px;
-  height: 48px;
-  padding: 0 20px;
-  border: none;
-  border: 1px solid transparent;
-  transition:
-    background 0.2s,
-    color 0.2s,
-    border-color 0.2s;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 16px;
-  line-height: 20px;
-  font-weight: 500;
-}
-
-a.primary {
-  background: var(--foreground);
-  color: var(--background);
-  gap: 8px;
-}
-
-a.secondary {
-  border-color: var(--gray-alpha-200);
-  min-width: 158px;
-}
-
-.footer {
-  grid-row-start: 3;
-  display: flex;
-  gap: 24px;
-}
-
-.footer a {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
-
-.footer img {
-  flex-shrink: 0;
-}
-
-/* Enable hover only on non-touch devices */
-@media (hover: hover) and (pointer: fine) {
-  a.primary:hover {
-    background: var(--button-primary-hover);
-    border-color: transparent;
-  }
-
-  a.secondary:hover {
-    background: var(--button-secondary-hover);
-    border-color: transparent;
-  }
-
-  .footer a:hover {
-    text-decoration: underline;
-    text-underline-offset: 4px;
-  }
-}
-
-@media (max-width: 600px) {
-  .page {
-    padding: 20px 10px;
-    padding-bottom: 80px;
-  }
-
-  .main {
-    align-items: center;
-  }
-
-  .main ol {
-    text-align: center;
-  }
-
-  .ctas {
-    flex-direction: column;
-  }
-
-  .ctas a {
-    font-size: 14px;
-    height: 40px;
-    padding: 0 16px;
-  }
-
-  a.secondary {
-    min-width: auto;
-  }
-
-  .footer {
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .userActions {
-    flex-direction: column;
-    width: 100%;
-  }
-
-  .primary, .secondary {
-    width: 100%;
-    text-align: center;
-  }
-
-  .pageItem {
-    flex-direction: column;
-    gap: 0.5rem;
-    align-items: flex-start;
-  }
-}
-
-@media (prefers-color-scheme: dark) {
-  .logo {
-    filter: invert();
-  }
-}
